pyproject.toml
[tool.poetry]
name = "coderecap"
version = "0.1.0"
description = "Repository Context Aware Source Code Summarization"
authors = ["Kanv Khare <kkhare@ncsu.edu>, Adithya Kiran <akiran2@ncsu.edu>, Sogolsadat Mansouri <smansou2@ncsu.edu>, Ziad Hassan <zmohame2@ncsu.edu>"]
license = "MIT License"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
streamlit = "^1.39.0"
langchain-ollama = "^0.2.0"
scikit-learn = "^1.5.2"
gitpython = "^3.1.43"
langchain = "^0.3.3"
langchain-community = "^0.3.5"
langchain-huggingface = "^0.1.2"
azure-ai-inference = "^1.0.0b5"
transformers = "^4.46.2"
faiss-cpu = "^1.9.0"
pytest = "^8.3.3"
pytest-cov = "^6.0.0"
python-dotenv = "^1.0.1"
datasets = "^3.1.0"
matplotlib = "^3.9.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
